% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_slash}
\alias{add_slash}
\title{add_slash}
\usage{
add_slash(path)
}
\arguments{
\item{path}{Character vector of length one; the path to which to add a slash.}
}
\value{
The path with a slash appended.
}
\description{
Add a trailing slash to a path.
}
\details{
Checks if a path has a trailing slash and adds one if not.
}
\examples{
add_slash("path/wihout/a/trailing_slash")
add_slash("path/already/has/a/trailing_slash/")
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
