% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{all_duplicated}
\alias{all_duplicated}
\title{all_duplicated}
\usage{
all_duplicated(x)
}
\arguments{
\item{x}{Vector or dataframe or array.}
}
\value{
For vectors, a logical vector indicating which elements are duplicated.
For dataframes, a logical vector indicating which rows are duplicated.
}
\description{
Find all duplicate values
}
\details{
Similar to base::duplicated(), but returns a logical vector
indicating all duplicate values (not just duplicates after
the first).
}
\examples{
all_duplicated(c("a","a","b"))
all_duplicated(data.frame(c("a","a","b", "d"), c("a","a","c", "e")))
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
