% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{blank_y_theme}
\alias{blank_y_theme}
\title{blank_y_theme}
\usage{
blank_y_theme()
}
\value{
A ggplot object
}
\description{
Remove y-axis text and ticks
}
\examples{
library(ggplot2)
p1 <- ggplot(iris, aes(Sepal.Length, Petal.Length)) +
  geom_point(aes(color = Species))
p1
p1 + blank_y_theme()

}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
