% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{blankplot}
\alias{blankplot}
\title{blankplot}
\usage{
blankplot()
}
\value{
A ggplot object that is completely blank
}
\description{
Make blank filler plot
}
\details{
Used for filling in multiplot grids when one of the grid squares should be blank.
}
\examples{
 library(ggplot)
 library(cowplot)
 p1 <- ggplot(iris, aes(Sepal.Length, Petal.Length)) +
   geom_point(aes(color = Species))
 p2 <- ggplot(iris, aes(Sepal.Length)) +
   geom_bar(stat="count")
 p3 <- ggplot(iris, aes(Sepal.Width, Petal.Width)) +
   geom_point(aes(color = Species))
 p4 <- blankplot()
 plot_grid(p1, p2, p3, p4, nrow = 2, ncol = 2)

}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
