% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_bib}
\alias{clean_bib}
\title{Clean up a latex bibliography file}
\usage{
clean_bib(raw_bib_file, max_line_length = 1000,
  strip_fields = c("abstract", "file", "keywords", "url", "doi", "issn",
  "isbn", "month", "number"), strip_mendeley_header = TRUE)
}
\arguments{
\item{raw_bib_file}{Path to raw .bib file; this function is designed in particular
to clean .bib files produced by Mendeley.}

\item{max_line_length}{Integer. Lines greater than this length will be removed from the
bibliography.}

\item{strip_fields}{Fields to remove from the bibliography. Note that "number"
is issue number.}

\item{strip_mendeley_header}{Logical; should the header added by Mendeley be removed?}
}
\value{
Character vector.
}
\description{
In particular one exported from Mendeley
}
\examples{
\dontrun{
# Clean bibliography.
clean_bib <- clean_bib("mendeley_refs.bib")
# Write out cleaned bibliography.
write_lines(clean_bib, "references.bib")
}
}
