% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{clean_legend_theme}
\alias{clean_legend_theme}
\title{clean_legend_theme}
\usage{
clean_legend_theme()
}
\value{
A ggplot object with legend entry boxes removed
}
\description{
Clean up ggplot legends
}
\details{
Removes the legend title and default boxes around ggplot legend entries.
}
\examples{
library(ggplot2)
p1 <- ggplot(iris, aes(Sepal.Length, Petal.Length)) +
  geom_point(aes(color = Species))
p1
p1 + clean_legend_theme()

 @export
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
