% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{english2}
\alias{english2}
\title{Spell out a number in English if it's less than 10}
\usage{
english2(x, limit = 10)
}
\arguments{
\item{x}{Number.}

\item{limit}{Upper limit of numbers to spell out
(default = 10).}
}
\value{
Character vector of length 1.
}
\description{
Spell out a number in English if it's less than 10
}
\examples{
english2(2)
english2(12)
english2(10)
}
