% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{fasttree}
\alias{fasttree}
\title{Infer a phylogenetic tree using FastTree}
\usage{
fasttree(
  seqs,
  mol_type = "dna",
  model = "gtr",
  gamma = FALSE,
  other_args = NULL,
  echo = FALSE,
  ...
)
}
\arguments{
\item{seqs}{DNA alignment of class DNAbin}

\item{mol_type}{Molecule type; either "dna" or "aa" (proteins)}

\item{model}{Model to use for phylogenetic analysis. Choose
"wag" (WAG+CAT), "lg" (LG+CAT), "gtr" (GTR+CAT), "jc" (Jukes-Cantor + CAT)}

\item{gamma}{Logical; should branch lengths be rescaled and Gamma20-based
likelihood calculated?}

\item{other_args}{Other arguments to pass to fasttree; must be entered as a character
vector with the name and value of each argument separately.}

\item{echo}{Logical; should STDERR and STDOUT be printed to the screen?}

\item{...}{Additional arguments; not used by this function.}
}
\value{
List of class "phylo".
}
\description{
Required FastTree to be installed and on user's PATH.
}
\examples{
\dontrun{
library(ape)
data(woodmouse)
fasttree(woodmouse)
}
}
\references{
http://www.microbesonline.org/fasttree/
}
