% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{flatten_DNA_list}
\alias{flatten_DNA_list}
\title{flatten_DNA_list}
\usage{
flatten_DNA_list(dna_list)
}
\arguments{
\item{dna_list}{List of objects of class \code{"DNAbin"}.}
}
\value{
Single object of class \code{"DNAbin"}.
}
\description{
Flatten a list of lists of DNA sequences into a list of DNA sequences
}
\details{
This comes in handy if \code{ape::read.FASTA()} is mapped onto
a vector of fasta files, where each fasta file may contain multiple
sequences.
}
\examples{
cat(">seq1",
    "ATCG",
    ">seq2",
    "GATC",
    ">seq2",
    "ATTG",
   file = tempfile("fasta1", fileext = ".fasta"), sep = "\\n")

cat(">seq4",
    "ATCG",
    ">seq5",
    "GATC",
    ">seq6",
    "ATTG",
   file = tempfile("fasta2", fileext = ".fasta"), sep = "\\n")

files <- list.files(tempdir(), pattern = "fasta", full.names = TRUE)

nested_seqs <- lapply(files, ape::read.FASTA)

flat_seqs <- flatten_DNA_list(nested_seqs)
}
