% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{The number of desired colors}
}
\value{
Vector of color codes
}
\description{
Produce ggplot-like colors
}
\details{
Shamelessly copied from \url{https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
}
\examples{
gg_color_hue(5)

}
