% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{iqtree}
\alias{iqtree}
\title{Run IQ-TREE}
\usage{
iqtree(
  alignment = NULL,
  wd = getwd(),
  aln_path = NULL,
  tree_path = NULL,
  bb = NULL,
  nt = NULL,
  alrt = NULL,
  m = NULL,
  redo = FALSE,
  spp = NULL,
  seed = NULL,
  echo = FALSE,
  other_args = NULL,
  ...
)
}
\arguments{
\item{alignment}{DNA alignment to use for phylogenetic analysis. Must be matrix
(i.e., aligned sequences) of class DNAbin}

\item{wd}{Path to working directory. The alignment and IQ-TREE intermediate files
and results will be written here.}

\item{aln_path}{Path to DNA alignment. Either alignment, or aln_path must be provided,
but not both}

\item{tree_path}{Optional; path to tree when it is written out by IQ-TREE, useful
if this differs from default alignment name.}

\item{bb}{Optional; number of ultrafast bootstrap replicates to run.}

\item{nt}{Optional; number of cores to use. Set to "AUTO" to determine automatically.}

\item{alrt}{Optional; number of SH-aLRT tests to run.}

\item{m}{Optional; specify model. If no model is given, ModelTest will be run
to identify the best model for the data.}

\item{redo}{Logical; should the analysis be redone from scratch if output from
previous runs is present?}

\item{spp}{Path to partition file.}

\item{seed}{Optional; Specify a random number seed to reproduce a previous run.}

\item{echo}{Logical; should STDERR be written to the screen?}

\item{other_args}{Other arguments to pass to IQ tree; must be entered as a character
vector with the name and value of each argument separately. For example,
c("-pers", "0.2", "-nstop", "500").}

\item{...}{Additional arguments; not used by this function.}
}
\value{
Phylogenetic tree (list of class "phylo")
}
\description{
For details, see http://www.iqtree.org/doc/
}
\examples{
\dontrun{
library(ape)
data(woodmouse)
# Rapid boot-strap tree with 1000 replicates on best-fitting model
tree <- iqtree(woodmouse, tempdir(), bb = 1000, echo = TRUE)
plot(tree)
# Check the optimum number of cores to use for GTR+I+G model
iqtree(tempdir(), woodmouse, m = "GTR+I+G", nt = "AUTO", echo = TRUE, redo = TRUE)
}
}
