% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_ref_list}
\alias{make_ref_list}
\title{Produce a filtered, cleaned .bib file for an Rmd document}
\usage{
make_ref_list(rmd_file, raw_bib, final_bib, exclude = NULL, ...)
}
\arguments{
\item{rmd_file}{Path to Rmd file}

\item{raw_bib}{Path to raw bib file}

\item{final_bib}{Path to write cleaned, filtered bib file}

\item{exclude}{Optional character vector of reference IDs to exclude
from the final bibliography.}

\item{...}{Additional named arguments passed to \code{\link{clean_bib}}
that will be used when cleaning the raw bib file.}
}
\value{
Nothing
}
\description{
Takes a raw .bib file (e.g., one exported from
Mendeley with thousands of entries), filters it
to only the citations in the Rmd file, and cleans
the references so pandoc won't crash.
}
\details{
Note that individual entries in the cleaned bib file may still
need additional editing.
}
