% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste3}
\alias{paste3}
\title{paste3}
\usage{
paste3(..., sep = " ")
}
\description{
Paste while removing NAs
}
\details{
Removes NAs from pasted elements, but if ALL elements are NA, the result is NA.

Shamelessly copied from
\url{https://stackoverflow.com/questions/13673894/suppress-nas-in-paste}
}
\examples{
paste3(c("a", "b", "c", NA), c("A","B", NA, NA))
}
