% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{powerpoint_theme}
\alias{powerpoint_theme}
\title{powerpoint_theme}
\usage{
powerpoint_theme()
}
\value{
ggplot object with larger font sizes
}
\description{
Increase font size of titles for powerpoint
}
\examples{
p1 <- ggplot(iris, aes(Sepal.Length, Petal.Length)) +
 geom_point(aes(color = Species))
p1 + powerpoint_theme()

}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
