% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{preview_pal}
\alias{preview_pal}
\title{Preview color palette}
\usage{
preview_pal(colors)
}
\arguments{
\item{colors}{A vector of colors}
}
\value{
Plot of the colors with their names if provided
}
\description{
Preview color palette
}
\examples{
preview_pal(c("red", "green"))
}
