% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resolve_col}
\alias{resolve_col}
\title{resolve_col}
\usage{
resolve_col(col, collapse_chr = ", ")
}
\arguments{
\item{col}{List (supposed to be single column of a dataframe).}

\item{collapse_chr}{Character to use to separate values when pasting together elements
of items in the list that contain > 1 unique element.}
}
\value{
List containing a single element per list item.
}
\description{
Find all duplicate values
}
\details{
Resolve duplicates within a single column of a dataframe
that should only contain a unique value. To be used internally
with \code{\link{resolve_duplicates}}
}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
