% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resolve_df}
\alias{resolve_df}
\title{resolve_df}
\usage{
resolve_df(df, collapse_chr)
}
\arguments{
\item{df}{The dataframe to which \code{\link{resolve_col}} should be applied.}

\item{collapse_chr}{Character to use to separate values when pasting together elements
of items in the list that contain > 1 element.}
}
\value{
Dataframe
}
\description{
Resolve duplicates across a dataframe
}
\details{
To be used internally with \code{\link{resolve_duplicates}}
}
