% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resolve_duplicates}
\alias{resolve_duplicates}
\title{resolve_duplicates}
\usage{
resolve_duplicates(data, search_col, collapse_chr = ", ")
}
\arguments{
\item{data}{A dataframe}

\item{search_col}{Vector of column name(s) of the dataframe to check for
duplicate values.}

\item{collapse_chr}{Character used to separate duplicate values when combining cells}
}
\value{
A dataframe with duplicated rows removed and conflicting values pasted together.
}
\description{
Resolve duplicate values in a dataframe by combining rows.
}
\details{
Unlike base::unique(), which returns unique rows taking into account all columns,
resolve_duplicates will only check for duplicate values in the columns specified by
\code{search_col}. Conflicting values in columns not included in \code{search_col},
are resolved by pasting together separated by \code{collapse_chr}.
}
\examples{
df <- data.frame(list(C1 = c("a","a","b","c","c","d","d"),
                     C2 = c("b","d","e","f","g","h","j"),
                     C3 = c("h","h","i","j","k","e","e")),
                 stringsAsFactors = FALSE)

unique(df) # doesn't do anything because df already unique
           # when using all columns
resolve_duplicates(df, "C1")
resolve_duplicates(df, "C3")
resolve_duplicates(df, c("C1", "C3"))
resolve_duplicates(df, "C2")
}
