% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{round_t}
\alias{round_t}
\title{Round a number and include trailing zeros}
\usage{
round_t(x, digits)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of digits for rounding.}
}
\value{
Character vector
}
\description{
Round a number and include trailing zeros
}
\examples{
round_t(c(2.12, 2.3), 4)
}
