% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{standard_theme}
\alias{standard_theme}
\title{standard_theme}
\usage{
standard_theme(font_size = 12)
}
\arguments{
\item{font_size}{Font size to use throughout}
}
\value{
A ggplot object
}
\description{
A simple ggplot theme
}
\details{
A simple ggplot theme with uniform font sizes and no extra gridlines.
}
\examples{
library(ggplot2)
p1 <- ggplot(iris, aes(Sepal.Length, Petal.Length)) +
  geom_point(aes(color = Species))
p1
p1 + standard_theme()

}
\author{
Joel H Nitta, \email{joelnitta@gmail.com}
}
